add set.go
